Fix schema errors
